perf: eliminate GPU cache thrashing for 2x speedup over MPS
Key optimizations:
- Single compiled function (CFG only) instead of two — avoids catastrophic
  GPU cache thrashing on unified memory that caused 30s switching cost per
  CFG→cond transition. Uncompiled B=1 matches compiled speed.
- Precompute RoPE cos/sin in prepare_inputs — eliminates 72 trig
  computations per denoising step (2 per layer × 36 layers).
- Free text encoder after encoding with gc.collect() — recovers ~7GB GPU
  memory, preventing swap thrashing during denoising.
- Materialize latents before denoising loop to avoid compiled fn retrace.
- Simplified AdaLN modulation (removed unnecessary float32 upcasts).

Benchmarks @ 1024×1024, 20 steps, cfg_cutoff=0.5:
- Denoising: 9.4-10.0s/step → ~2.0x vs PyTorch MPS (19s/step)
- Isolated (no text encoder overhead): 8.8-9.3s/step → 2.0-2.15x
- Active memory: 27-43GB (down from 64GB with dual compiled fns)

+++ ernie_mlx/model/dit_block.py
@@ -41,18 +41,15 @@ class DiTBlock(nn.Module):
         # Self-attention branch
         residual = x
         x_norm = self.adaLN_sa_ln(x)
-        # AdaLN modulation in float32
-        x_mod = (x_norm.astype(mx.float32) * (1 + scale_msa.astype(mx.float32)) + shift_msa.astype(mx.float32))
-        x_mod = x_mod.astype(x.dtype)
+        x_mod = x_norm * (1 + scale_msa) + shift_msa
         attn_out = self.self_attention(x_mod, mask=attention_mask, rotary_pos_emb=rotary_pos_emb)
-        x = residual + (gate_msa.astype(mx.float32) * attn_out.astype(mx.float32)).astype(x.dtype)
+        x = residual + gate_msa * attn_out
 
         # FFN branch
         residual = x
         x_norm = self.adaLN_mlp_ln(x)
-        x_mod = (x_norm.astype(mx.float32) * (1 + scale_mlp.astype(mx.float32)) + shift_mlp.astype(mx.float32))
-        x_mod = x_mod.astype(x.dtype)
+        x_mod = x_norm * (1 + scale_mlp) + shift_mlp
         ffn_out = self.mlp(x_mod)
-        x = residual + (gate_mlp.astype(mx.float32) * ffn_out.astype(mx.float32)).astype(x.dtype)
+        x = residual + gate_mlp * ffn_out
 
         return x
